Dropped preparing html GUI
Not efficient work. Already learned the basics of flask with stuff made until this moment.

It would take a lot of time to prepare messages for user in GUI in way like they are working in terminal.
Because of terminal way is working pretty well, i'm leaving as it is.